requirements.txt
pyserial>=3.5
openai>=1.0.0
python-dotenv>=1.0.0

